refactor: user journey
