+++ components/WhiteListConditions.js
@@ -3,12 +3,14 @@ import { useRouter } from 'next/router';
 
 import { useContext } from 'react';
 
+import PropTypes from 'prop-types';
+
 import { useAuth } from '../contexts/auth';
 import { WalletDispatch, WalletState } from '../contexts/wallet';
 import { __ } from '../utils/helpers';
 import ProfilePhoto from './ProfilePhoto';
 
-const WhiteListConditions = ({ referredBy }) => {
+const WhiteListConditions = ({ referredByUser }) => {
   const router = useRouter();
   const { isAuthenticated } = useAuth();
   const { connectWallet, switchNetwork } = useContext(WalletDispatch);
@@ -16,18 +18,20 @@ const WhiteListConditions = ({ referredBy }) => {
     useContext(WalletState);
   return (
     <>
-      <h2 className="mt-8 text-3xl md:text-4xl leading-snug md:items-center flex flex-col md:flex-row">
-        <span>{__('token_sale_invite_page_invited_by')}</span>
-        <div className="w-fit bg-primary-light flex items-center ml-2 gap-2 px-4 py-2">
-          <ProfilePhoto user={referredBy} size="sm" />
-          <span>{' ' + referredBy?.screenname}</span>
-        </div>
-      </h2>
+      {referredByUser && (
+        <h2 className="mt-8 text-3xl md:text-4xl leading-snug md:items-center flex flex-col md:flex-row">
+          <span>{__('token_sale_invite_page_invited_by')}</span>
+          <div className="w-fit bg-primary-light flex items-center ml-2 gap-2 px-4 py-2">
+            <ProfilePhoto user={referredByUser} size="sm" />
+            <span>{' ' + referredByUser?.screenname}</span>
+          </div>
+        </h2>
+      )}
       <h3 className="mt-4 text-4xl leading-snug">
         {__('token_sale_invite_page_condition_title')}
       </h3>
       <ol className="list-decimal pl-8 pb-8 md:pb-0">
-        <li className="mt-4 text-xl md:text-2xl leading-snug">
+        <li className="mt-4 text-xl md:text-2xl leading-snug relative">
           <Link href={`/signup?back=${router.asPath}`} passHref>
             <a className="text-primary underline cursor-pointer">
               {__('signup_form_create')}
@@ -43,7 +47,7 @@ const WhiteListConditions = ({ referredBy }) => {
             <img
               src="/images/token-sale/circle-check.svg"
               alt=""
-              className="inline w-8 md:w-12 ml-5"
+              className="inline w-8 md:w-12 ml-5 absolute -top-1"
             />
           )}
         </li>
@@ -78,6 +82,16 @@ const WhiteListConditions = ({ referredBy }) => {
       <p>{__('token_sale_invite_page_process_price')}</p>
     </>
   );
+};
+
+WhiteListConditions.propTypes = {
+  referredByUser: PropTypes.oneOf([
+    PropTypes.shape({
+      _id: PropTypes.string,
+      screenname: PropTypes.string,
+    }),
+    undefined,
+  ]),
 };
 
 export default WhiteListConditions;
