+++ pages/login/index.js
@@ -79,74 +79,69 @@ const Login = () => {
       <Head>
         <title>{__('login_title')}</title>
       </Head>
-      <div className="mural">
-        <main className="main-content max-w-prose center intro">
-          <form onSubmit={onSubmit}>
-            <div className="w-full mb-4">
-              <label
-                className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
-                htmlFor="email"
-              >
-                {__('login_email')}
-              </label>
-              <input
-                className="w-full bg-transparent"
-                type="email"
-                name="email"
-                id="email"
-                value={email}
-                placeholder="name@awesomeproject.co"
-                onChange={(e) => setEmail(e.target.value)}
-                required
-              />
-            </div>
-            <div className="w-full mb-4">
-              <label
-                className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
-                htmlFor="password"
-              >
-                {__('login_password')}
-              </label>
-              <input
-                className="w-full bg-transparent"
-                type="password"
-                name="password"
-                id="password"
-                value={password}
-                placeholder="*****"
-                onChange={(e) => setPassword(e.target.value)}
-                required
-              />
-            </div>
-            <div className="card-footer">
-              <div className="flex flex-col justify-between items-center gap-4 sm:flex-row">
-                <div className="flex flex-col gap-4 w-full sm:flex-row">
-                  <button
-                    type="submit"
-                    className="btn-primary w-full sm:w-auto"
-                  >
-                    {__('login_submit')}
-                  </button>
-                  <button
-                    type="submit"
-                    className="btn-primary"
-                    onClick={walletConnectAndSignInFlow}
-                  >
-                    {__('blockchain_sign_in_with_wallet')}
-                  </button>
-                </div>
-
-                <Link href="/login/forgot-password" as="/login/forgot-password">
-                  <a className="whitespace-nowrap">
-                    {__('login_link_forgot_password')}
-                  </a>
-                </Link>
+      <main className="main-content max-w-prose center intro flex items-center flex-1">
+        <form onSubmit={onSubmit}>
+          <div className="w-full mb-4">
+            <label
+              className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+              htmlFor="email"
+            >
+              {__('login_email')}
+            </label>
+            <input
+              className="w-full bg-transparent"
+              type="email"
+              name="email"
+              id="email"
+              value={email}
+              placeholder="name@awesomeproject.co"
+              onChange={(e) => setEmail(e.target.value)}
+              required
+            />
+          </div>
+          <div className="w-full mb-4">
+            <label
+              className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+              htmlFor="password"
+            >
+              {__('login_password')}
+            </label>
+            <input
+              className="w-full bg-transparent"
+              type="password"
+              name="password"
+              id="password"
+              value={password}
+              placeholder="*****"
+              onChange={(e) => setPassword(e.target.value)}
+              required
+            />
+          </div>
+          <div className="card-footer">
+            <div className="flex flex-col justify-between items-center gap-4 sm:flex-row">
+              <div className="flex flex-col gap-4 w-full sm:flex-row">
+                <button type="submit" className="btn-primary w-full sm:w-auto">
+                  {__('login_submit')}
+                </button>
+                <button
+                  type="submit"
+                  className="btn-primary"
+                  onClick={walletConnectAndSignInFlow}
+                >
+                  {__('blockchain_sign_in_with_wallet')}
+                </button>
               </div>
+
+              <Link href="/login/forgot-password" as="/login/forgot-password">
+                <a className="whitespace-nowrap">
+                  {__('login_link_forgot_password')}
+                </a>
+              </Link>
             </div>
-          </form>
-          {error && <p className="text-primary mt-4 text-center">{error}</p>}
-        </main>
-      </div>
+          </div>
+        </form>
+        {error && <p className="text-primary mt-4 text-center">{error}</p>}
+      </main>
     </Layout>
   );
 };
