--- utils/temp.js
@@ -1,2 +1,0 @@
-const arr = [1, 2, 3, { year: 2023 }];
-console.log(new Map(arr));
